# clients/rust/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@flowcatalyst.tech
    name: FlowCatalyst Support
    url: https://flowcatalyst.tech/support
  description: Control plane API for FlowCatalyst - multi-tenant event routing platform.
  license:
    name: Proprietary
    url: https://flowcatalyst.tech/license
  title: FlowCatalyst Platform API
  version: "1.0"
paths:
  /admin/platform/anchor-domains:
    get:
      description: Returns all anchor domains that grant platform-wide access
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/internal_platform_api.AnchorDomainDTO"
                type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List anchor domains
      tags:
        - Anchor Domains
    post:
      description: Adds a new anchor domain that grants platform-wide access to users
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_platform_api.CreateAnchorDomainRequest"
        description: Anchor domain to add
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.AnchorDomainDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Add anchor domain
      tags:
        - Anchor Domains
...
  /admin/platform/audit-logs:
    get:
      description: Returns audit logs with optional filtering by entity type, entity
        ID, principal, or operation
      parameters:
        - description: Filter by entity type
          in: query
          name: entityType
          schema:
            type: string
        - description: Filter by entity ID
          in: query
          name: entityId
          schema:
            type: string
        - description: Filter by principal ID
          in: query
          name: principalId
          schema:
            type: string
        - description: Filter by operation name
          in: query
          name: operation
          schema:
            type: string
        - description: Page number (0-based)
          in: query
          name: page
          schema:
            type: integer
            default: 0
        - description: Page size
          in: query
          name: pageSize
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_audit.AuditL\
                  ogListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List audit logs
      tags:
        - Audit Logs
...
  /admin/platform/auth-configs:
    get:
      description: Returns all authentication configurations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
                type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List auth configs
      tags:
        - Auth Config
    post:
      description: Creates a new authentication configuration for an email domain
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_platform_api.CreateAuthConfigRequest"
        description: Auth config to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create auth config
      tags:
        - Auth Config
...
  /api/admin/platform/roles:
    get:
      description: Returns a list of all roles in the system
      parameters:
        - description: Filter by scope (ANCHOR, PARTNER, CLIENT)
          in: query
          name: scope
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_role.Role"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List all roles
      tags:
        - Admin - Roles
    post:
      description: Creates a new role with specified permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_role_operati\
                ons.CreateRoleCommand"
        description: Role details
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_role.Role"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "409":
          description: Role with code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create a new role
      tags:
        - Admin - Roles
...
  /api/admin/platform/service-accounts:
    get:
      description: Returns a list of all service accounts in the system
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_serviceaccou\
                    nt.ServiceAccount"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List all service accounts
      tags:
        - Admin - Service Accounts
    post:
      description: Creates a new service account for machine-to-machine
        authentication. Returns credentials only on creation.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_serviceaccou\
                nt_operations.CreateServiceAccountCommand"
        description: Service account details
        required: true
      responses:
        "201":
          description: Created service account with credentials (secret shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/go_flowcatalyst_tech_internal_platform_serviceaccou\
                  nt.ServiceAccount"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "409":
          description: Service account with name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create a new service account
      tags:
        - Admin - Service Accounts
...
  /bff/debug/dispatch-jobs:
    get:
      description: List raw dispatch jobs from the transactional collection
        (debug/admin only)
      parameters:
        - description: Page number (0-based)
          in: query
          name: page
          schema:
            type: integer
            default: 0
        - description: Page size (max 100)
          in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.PagedRawDispatchJobResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List raw dispatch jobs
      tags:
        - Raw Dispatch Jobs (Debug)
...
servers:
  - url: //localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer {token}"'
      in: header
      name: Authorization
      type: apiKey
    SessionCookie:
      description: Session cookie for browser-based authentication
      in: cookie
      name: FLOWCATALYST_SESSION
      type: apiKey
  schemas:
...
    internal_platform_api.AnchorDomainDTO:
      properties:
        createdAt:
          type: string
        domain:
          type: string
        id:
          type: string
      type: object
    internal_platform_api.AuthConfigDTO:
      properties:
        additionalClientIds:
          items:
            type: string
          type: array
        authProvider:
          type: string
        configType:
          type: string
        createdAt:
          type: string
        emailDomain:
          type: string
        entraTenantId:
          type: string
        grantedClientIds:
          items:
            type: string
          type: array
        groupsClaim:
          type: string
        id:
          type: string
        idpType:
          type: string
        oidcClientId:
          type: string
        oidcIssuerPattern:
          type: string
        oidcIssuerUrl:
          type: string
        oidcMultiTenant:
          type: boolean
        primaryClientId:
          type: string
        rolesClaim:
          type: string
        updatedAt:
          type: string
      type: object
...
    internal_platform_api.CreateAnchorDomainRequest:
      properties:
        domain:
          type: string
      type: object
...
    internal_platform_api.ErrorResponse:
      properties:
        details:
          type: string
        error:
          type: string
        message:
          type: string
      type: object
...